(GH-23) Update to get actual item before removing in list
Due to powershell and C# code working a little differently in some cases,
there are some times a problem when using powershell to check if a list
contains an item, and then using C# to remove the actual item.
This is because a cast between a string and an object happens in C#
which makes it think the items are no longer equal and causes an
infinite loop.
This commit makes the necessary change to get the actual item in the
list before trying to remove the item, so no casting is automatically
performed.